move post model
